fix for issue #430 where " or ' exist as valid data adjacent right to delimiter
In some cases data conains leading quotes formed by " or ' and are right adjacent to the delimiter.  In thes cases we want to treat all data between commas as data including those characters.  So we need a way to tell papaparse to ignore quotes.  However papaparse takes all false statements and defaults the quoteChar field to " which creates issues parsing the data when the above data example is present.