Fast headerMap to find duplicated headers
The current implementation used to rely on an instance of `Array` and on its `includes` method. Depending on how many headers one may have to export in the CSV this `Array` can contain many entries and performings `includes` on it might be costly as it grows (most of the cases will have to iterate over the whole collection as duplicates is rather the exception than the rule).

The current PR proposes to rely on a native Set (when available) or a simple polyfill for it. (Side note: I did not know how to put tests on the polyfill as the class is internal and never exposed so I tested it manually on my side)

Please feel free to close this PR if it does not make sense for the project. And thanks a lot for papaparse!